FixedThreadPool directly extends ThreadPoolExecutor.
